Create GitHubHttpUrlConnectionClient to encapsulate interactions with HttpUrlConnection
